Move utils into scripts folder (for inclusion in ejection)
