+++ packages/react-scripts/scripts/utils/createWebpackCompiler.js
